Add PaneList (GraphPane collection); Add PaneBase as parent of GraphPane and MasterPane

+++ source/ZedGraph/BarItem.cs
@@ -29,7 +29,7 @@ namespace ZedGraph
 	/// Summary description for BarItem.
 	/// </summary>
 	/// <author> John Champion </author>
-	/// <version> $Revision: 3.9 $ $Date: 2005-01-08 08:28:07 $ </version>
+	/// <version> $Revision: 3.10 $ $Date: 2005-01-22 06:20:49 $ </version>
 	[Serializable]
 	public class BarItem : CurveItem, ICloneable, ISerializable
 	{
@@ -168,10 +168,10 @@ namespace ZedGraph
 		/// <param name="scaleFactor">
 		/// The scaling factor to be used for rendering objects.  This is calculated and
 		/// passed down by the parent <see cref="ZedGraph.GraphPane"/> object using the
-		/// <see cref="GraphPane.CalcScaleFactor"/> method, and is used to proportionally adjust
+		/// <see cref="PaneBase.CalcScaleFactor"/> method, and is used to proportionally adjust
 		/// font sizes, etc. according to the actual size of the graph.
 		/// </param>
-		override public void Draw( Graphics g, GraphPane pane, int pos, double scaleFactor  )
+		override public void Draw( Graphics g, GraphPane pane, int pos, float scaleFactor  )
 		{
 			// Pass the drawing onto the bar class
 			if ( this.isVisible )
@@ -195,10 +195,10 @@ namespace ZedGraph
 		/// <param name="scaleFactor">
 		/// The scaling factor to be used for rendering objects.  This is calculated and
 		/// passed down by the parent <see cref="ZedGraph.GraphPane"/> object using the
-		/// <see cref="GraphPane.CalcScaleFactor"/> method, and is used to proportionally adjust
+		/// <see cref="PaneBase.CalcScaleFactor"/> method, and is used to proportionally adjust
 		/// font sizes, etc. according to the actual size of the graph.
 		/// </param>
-		override public void DrawLegendKey( Graphics g, GraphPane pane, RectangleF rect, double scaleFactor )
+		override public void DrawLegendKey( Graphics g, GraphPane pane, RectangleF rect, float scaleFactor )
 		{
 			this.bar.Draw( g, pane, rect, scaleFactor, true );
 		}
