+++ source/ZedGraph/CurveList.cs
@@ -30,7 +30,7 @@ namespace ZedGraph
 	/// 
 	/// <author> John Champion
 	/// modified by Jerry Vos</author>
-	/// <version> $Revision: 3.20 $ $Date: 2005-01-19 05:54:52 $ </version>
+	/// <version> $Revision: 3.21 $ $Date: 2005-01-22 06:20:50 $ </version>
 	[Serializable]
 	public class CurveList : CollectionPlus, ICloneable
 	{
@@ -534,10 +534,10 @@ namespace ZedGraph
 		/// <param name="scaleFactor">
 		/// The scaling factor to be used for rendering objects.  This is calculated and
 		/// passed down by the parent <see cref="GraphPane"/> object using the
-		/// <see cref="GraphPane.CalcScaleFactor"/> method, and is used to proportionally adjust
+		/// <see cref="PaneBase.CalcScaleFactor"/> method, and is used to proportionally adjust
 		/// font sizes, etc. according to the actual size of the graph.
 		/// </param>
-		public void Draw( Graphics g, GraphPane pane, double scaleFactor )
+		public void Draw( Graphics g, GraphPane pane, float scaleFactor )
 		{
 			// Configure the accumulator for stacked bars
 			//Bar.ResetBarStack();
